@@ -1,12 +1,11 @@
 import logging
 import random
-import sys
 from collections import defaultdict
 from enum import Enum
 from functools import reduce
 from math import inf
 from pathlib import Path
-from typing import Dict, List, Optional, Union
+from typing import Dict, List, Literal, Optional, Union
 
 from scipy.stats import pearsonr, spearmanr
 from sklearn.metrics import mean_absolute_error, mean_squared_error
@@ -16,6 +15,7 @@ from torch.utils.data import Dataset
 import flair
 from flair.data import DT, Dictionary, Sentence, _iter_dataset
 
+EmbeddingStorageMode = Literal["none", "cpu", "gpu"]
 log = logging.getLogger("flair")
 
 
@@ -24,10 +24,11 @@ class Result:
         self,
         main_score: float,
         detailed_results: str,
-        classification_report: dict = {},
-        scores: dict = {},
+        classification_report: Optional[dict] = None,
+        scores: Optional[Dict] = None,
     ) -> None:
-        assert "loss" in scores, "No loss provided."
+        classification_report = classification_report if classification_report is not None else {}
+        assert scores is not None and "loss" in scores, "No loss provided."
 
         self.main_score: float = main_score
         self.scores = scores
@@ -80,12 +81,7 @@ class MetricRegression:
         return "\t_\t_\t_\t_"
 
     def __str__(self) -> str:
-        line = "mean squared error: {:.4f} - mean absolute error: {:.4f} - pearson: {:.4f} - spearman: {:.4f}".format(
-            self.mean_squared_error(),
-            self.mean_absolute_error(),
-            self.pearsonr(),
-            self.spearmanr(),
-        )
+        line = f"mean squared error: {self.mean_squared_error():.4f} - mean absolute error: {self.mean_absolute_error():.4f} - pearson: {self.pearsonr():.4f} - spearman: {self.spearmanr():.4f}"
         return line
 
 
@@ -355,10 +351,7 @@ def convert_labels_to_one_hot(label_list: List[List[str]], label_dict: Dictionar
 
 
 def log_line(log):
-    if sys.version_info >= (3, 8):
-        log.info("-" * 100, stacklevel=3)
-    else:
-        log.info("-" * 100)
+    log.info("-" * 100, stacklevel=3)
 
 
 def add_file_handler(log, output_file):
@@ -372,7 +365,9 @@ def add_file_handler(log, output_file):
 
 
 def store_embeddings(
-    data_points: Union[List[DT], Dataset], storage_mode: str, dynamic_embeddings: Optional[List[str]] = None
+    data_points: Union[List[DT], Dataset],
+    storage_mode: EmbeddingStorageMode,
+    dynamic_embeddings: Optional[List[str]] = None,
 ):
     if isinstance(data_points, Dataset):
         data_points = list(_iter_dataset(data_points))
@@ -396,7 +391,7 @@ def store_embeddings(
             data_point.to("cpu", pin_memory=pin_memory)
 
 
-def identify_dynamic_embeddings(data_points: List[DT]):
+def identify_dynamic_embeddings(data_points: List[DT]) -> Optional[List]:
     dynamic_embeddings = []
     all_embeddings = []
     for data_point in data_points:
